requirements.txt
edge-tts
Pillow